# pkg/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  files.File:
    properties:
      created:
        type: string
      id:
        type: integer
      mime:
        type: string
      name:
        type: string
      size:
        type: integer
    type: object
...
  models.Message:
    properties:
      message:
        description: A standard message.
        type: string
    type: object
...
  models.TaskAttachment:
    properties:
      created:
        type: string
      created_by:
        $ref: '#/definitions/user.User'
      file:
        $ref: '#/definitions/files.File'
      id:
        type: integer
      task_id:
        type: integer
    type: object
...
  user.User:
    properties:
      created:
        description: A timestamp when this task was created. You cannot change this
          value.
        type: string
      email:
        description: The user's email address.
        maxLength: 250
        type: string
      id:
        description: The unique, numeric id of this user.
        type: integer
      name:
        description: The full name of the user.
        type: string
      updated:
        description: A timestamp when this task was last updated. You cannot change
          this value.
        type: string
      username:
        description: The username of the user. Is always unique.
        maxLength: 250
        minLength: 1
        type: string
    type: object
...
  v1.UserAvatarProvider:
    properties:
      avatar_provider:
        description: The avatar provider. Valid types are `gravatar` (uses the user
          email), `upload`, `initials`, `marble` (generates a random avatar for each
          user), `ldap` (synced from LDAP server), `openid` (synced from OpenID provider),
          `default`.
        type: string
    type: object
...
  v1.UserSettings:
    properties:
      default_project_id:
        description: |-
          If a task is created without a specified project this value should be used. Applies
          to tasks made directly in API and from clients.
        type: integer
      discoverable_by_email:
        description: If true, the user can be found when searching for their exact
          email.
        type: boolean
      discoverable_by_name:
        description: If true, this user can be found by their name or parts of it
          when searching for it.
        type: boolean
      email_reminders_enabled:
        description: If enabled, sends email reminders of tasks to the user.
        type: boolean
      extra_settings_links:
        additionalProperties: {}
        description: Additional settings links as provided by openid
        type: object
      frontend_settings:
        description: Additional settings only used by the frontend
      language:
        description: The user's language
        type: string
      name:
        description: The new name of the current user.
        type: string
      overdue_tasks_reminders_enabled:
        description: If enabled, the user will get an email for their overdue tasks
          each morning.
        type: boolean
      overdue_tasks_reminders_time:
        description: The time when the daily summary of overdue tasks will be sent
          via email.
        type: string
      timezone:
        description: The user's time zone. Used to send task reminders in the time
          zone of the user.
        type: string
      week_start:
        description: The day when the week starts for this user. 0 = sunday, 1 = monday,
          etc.
        type: integer
    type: object
...
  web.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact:
    email: hello@vikunja.io
    name: General Vikunja contact
    url: https://vikunja.io/contact/
  description: |-
    # Pagination
    Every endpoint capable of pagination will return two headers:
    * `x-pagination-total-pages`: The total number of available pages for this request
    * `x-pagination-result-count`: The number of items returned for this request.
    # Permissions
    All endpoints which return a single item (project, task, etc.) - no array - will also return a `x-max-permission` header with the max permission the user has on this item as an int where `0` is `Read Only`, `1` is `Read & Write` and `2` is `Admin`.
    This can be used to show or hide ui elements based on the permissions the user has.
    # Errors
    All errors have an error code and a human-readable error message in addition to the http status code. You should always check for the status code in the response, not only the http status code.
    Due to limitations in the swagger library we're using for this document, only one error per http status code is documented here. Make sure to check the [error docs](https://vikunja.io/docs/errors/) in Vikunja's documentation for a full list of available error codes.
    # Authorization
    **JWT-Auth:** Main authorization method, used for most of the requests. Needs `Authorization: Bearer <jwt-token>`-header to authenticate successfully.

...

    **BasicAuth:** Only used when requesting tasks via CalDAV.
...
  license:
...
paths:
...
  /tasks/{id}/attachments:
    get:
      consumes:
      - application/json
      description: Get all task attachments for one task.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: The page number. Used for pagination. If not provided, the first
          page of results is returned.
        in: query
        name: page
        type: integer
      - description: The maximum number of items per page. Note this parameter is
          limited by the configured maximum of items per page.
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: All attachments for this task
          schema:
            items:
              $ref: '#/definitions/models.TaskAttachment'
            type: array
        "403":
          description: No access to this task.
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: The task does not exist.
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Get  all attachments for one task.
      tags:
      - task
    put:
      consumes:
      - multipart/form-data
      description: Upload a task attachment. You can pass multiple files with the
        files form param.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: The file, as multipart form file. You can pass multiple.
        in: formData
        name: files
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attachments were uploaded successfully.
          schema:
            $ref: '#/definitions/models.Message'
        "403":
          description: No access to the task.
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: The task does not exist.
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Upload a task attachment
      tags:
      - task
  /tasks/{id}/attachments/{attachmentID}:
    delete:
      consumes:
      - application/json
      description: Delete an attachment.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attachment ID
        in: path
        name: attachmentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: The attachment was deleted successfully.
          schema:
            $ref: '#/definitions/models.Message'
        "403":
          description: No access to this task.
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: The task does not exist.
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Delete an attachment
      tags:
      - task
    get:
      description: Get one attachment for download. **Returns json on error.**
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attachment ID
        in: path
        name: attachmentID
        required: true
        type: integer
      - description: The size of the preview image. Can be sm = 100px, md = 200px,
          lg = 400px or xl = 800px. If provided, a preview image will be returned
          if the attachment is an image.
        in: query
        name: preview_size
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: The attachment file.
          schema:
            type: file
        "403":
          description: No access to this task.
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: The task does not exist.
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Get one attachment.
      tags:
      - task
...
  /user/settings/avatar:
    get:
      consumes:
      - application/json
      description: Returns the current user's avatar setting.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserAvatarProvider'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Return user avatar setting
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Changes the user avatar. Valid types are gravatar (uses the user
        email), upload, initials, marble, ldap (synced from LDAP server), openid (synced
        from OpenID provider), default.
      parameters:
      - description: The user's avatar setting
        in: body
        name: avatar
        required: true
        schema:
          $ref: '#/definitions/v1.UserAvatarProvider'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Set the user's avatar
      tags:
      - user
...
  /user/settings/general:
    post:
      consumes:
      - application/json
      parameters:
      - description: The updated user settings
        in: body
        name: avatar
        required: true
        schema:
          $ref: '#/definitions/v1.UserSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Change general user settings of the current user.
      tags:
      - user
...
  /user/timezones:
    get:
      consumes:
      - application/json
      description: Because available time zones depend on the system Vikunja is running
        on, this endpoint returns a project of all valid time zones this particular
        Vikunja instance can handle. The project of time zones is not sorted, you
        should sort it on the client.
      produces:
      - application/json
      responses:
        "200":
          description: All available time zones.
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Get all available time zones on this vikunja instance
      tags:
      - user
...
securityDefinitions:
  BasicAuth:
    type: basic
  JWTKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"